requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
redis
asgiref
requests
aioredis
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.1.0
casbin
paho-mqtt
python-multipart
celerybeat-mongo
pydantic-settings
pymongo[srv]
dnspython
celery[beat]
gmqtt
python-dotenv
pydantic
httpx
